create mpnet with weighted cross loss

+++ train-mpnet.sh
@@ -1,13 +1,13 @@
 echo "start training"
 git_status=$(git status -s) 
 
-# if [ -n "$git_status" ]; then
-#     echo "You have uncommitted changes. Please commit them first."
-#     exit 1
-# fi
+if [ -n "$git_status" ]; then
+    echo "You have uncommitted changes. Please commit them first."
+    exit 1
+fi
 
-training_script=core.train_scripts.huggingface_training
-training_logs=./training_logs/huggingface/
+training_script=core.train_scripts.mpnet_training
+training_logs=./training_logs/mpnet/
 export WANDB_PROJECT=focus_persona_classification
 # clear dir
 rm -rf $training_logs*
